Update TorchSeq2PC.py
Fixed sign error in definition of epsilon inside StrictPCPredErrs function. The sign error was inconsequential for training the network (updating parameters), but it resulted in epsilon using a different convention for the Strict versus FixedPrediction cases. Essentially, epsilon was defined as the negative prediction error instead of the positive prediction error.